ReadNumber: float st-temps eliminate 8-byte align prologue (-55 byte_diffs)
flow between float values causes MSVC6 to allocate a 64-bit slot,
which forces the function frame to 8-byte alignment via the
'mov ebp,esp; and esp,0xfffffff8' prologue — absent in the
reference binary.

token.floatvalue (float) or (double)(int) values immediately
compared against fielddef_float (float) results, so single
precision matches the original semantics. MSVC keeps the chain
on x87 either way; only the spill type changes.

ReadNumber: OUR+3 (was OUR+20), 1476 byte_diffs (was 1531).

Source-commit: 4946b5d26458a398f9af557d25af28e3bf266efe

+++ botlib/botlib.c
@@ -38074,15 +38074,15 @@ int __cdecl ReadNumber(source_t *src, char **field, float *out)
   int v3; // esi
   int result; // eax
   int v5; // eax
-  double v6; // st7
+  float v6; // st7
   int v7; // ebx
   int v8; // ecx
   int v9; // esi
   int v10; // rax
-  double v11; // st7
-  double v12; // st7
+  float v11; // st7
+  float v12; // st7
   int v13; // rax
-  double v14; // st7
+  float v14; // st7
   char v15; // [esp+10h] [ebp-450h]
   int v16; // [esp+20h] [ebp-440h]
   int v17; // [esp+24h] [ebp-43Ch]
